refactor: renamed the `user_vault.py` schema file to follow PEP 8 module file naming convention instead of class naming convention.
